+++ lib/components/text_input_field.dart
@@ -1,24 +1,28 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:ui_bits/components/components.dart';
+import 'package:ui_bits/internal_ui_bits.dart';
 
-class BitTextInputField extends StatelessWidget {
+class BitInputTextField extends StatelessWidget {
   final FieldLabels messages;
   final Field<String> field;
+  final BitAnimation animation;
 
-  const BitTextInputField(
+  const BitInputTextField(
     this.messages, {
     this.field,
+    this.animation = const BitNoAnimation(),
   });
 
   @override
   Widget build(BuildContext context) {
-    return TextFormField(
-      controller: field?.controller,
-      keyboardType: TextInputType.visiblePassword,
-      decoration: InputDecoration(
-        labelText: messages.label,
-        prefixIcon: BitInputFieldIcon(messages.icon),
+    return animation.wrapWidget(
+      child: TextFormField(
+        controller: field?.controller,
+        keyboardType: TextInputType.visiblePassword,
+        decoration: InputDecoration(
+          labelText: messages.label,
+          prefixIcon: BitInputFieldIcon(messages.icon),
+        ),
       ),
     );
   }