+++ lib/ui_bits.dart
@@ -1,5 +1,6 @@
 library ui_bits;
 
+export 'animated_components/animated_components.dart';
 export 'animations/animations.dart';
 export 'components/components.dart';
 export 'themes/theme_factory.dart';